# doc/03_openapi_koipa_kl.yaml
openapi: 3.0.3
info:
  title: Koipa AI Engine API
  description: |
    한국지식재산보호원 AI 영업비밀관리시스템 — 한국지식재산보호원 AI 엔진의
    KL 웹시스템 연동용 REST API 명세.

    담당 기능: FUN-003, FUN-004, FUN-005, FUN-022, FUN-023, FUN-024
    호출 주체: KL 웹시스템 (Backend)
    제공 주체: Koipa AI Engine (FastAPI)

    인증: API Key (서버 간 통신) + 선택적 JWT (관리자 행위 추적)
    모든 응답은 application/json. 파일 업로드는 multipart/form-data.
    날짜·시각은 ISO 8601 UTC. 모든 ID는 UUIDv4.

    내부 RAG 백엔드: PostgreSQL pgvector dense + bigram tsvector `ts_rank`
    하이브리드. 본 명세는 backend-agnostic 하므로 KL 측은 RAG 동작 세부를
    몰라도 호출 가능. ES 경로는 레거시 어댑터로만 남아 있다.
  version: 1.0.0-draft
  contact:
    name: Koipa AI Engineering
servers:
  - url: http://koipa-api:8000/api/v1
    description: 온프레미스 Docker 내부망
  - url: https://{host}/api/v1
    description: 프로덕션
    variables:
      host:
        default: ai.example.koipa.local
security:
  - ApiKeyAuth: []
...
paths:
...
  /documents:
    post:
      tags: [documents]
      summary: 분류 대상 문서 업로드·ingestion
      description: |
        등급 판정을 받을 업무 문서(HWP/PDF/DOCX/TXT 등)를 업로드한다.
        포맷 자동 감지 → 텍스트 추출(필요 시 OCR) → 원본 스토리지 저장 →
        documents/chunks 테이블 적재. 이후 POST /classify?doc_id=<doc_id> 로 등급 요청.
        /guide/documents(지식베이스용 가이드 문서)와 완전 분리된 도메인.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [actor, file]
              properties:
                actor:
                  type: string
                  description: "Actor JSON 문자열 (user_id, role)"
                doc_type:
                  type: string
                  nullable: true
                external_ref:
                  type: string
                  nullable: true
                  description: "외부 문서 ID (EDMS 등)"
                file:
                  type: string
                  format: binary
                  description: "HWP / HWPX / PDF / DOCX / TXT"
      responses:
        '201':
          description: ingestion 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: "파일 크기 초과 (MAX_UPLOAD_MB)" }
        '422': { description: "빈 파일 또는 actor JSON 오류" }
        '500': { $ref: '#/components/responses/InternalError' }
  /documents/analyze:
    post:
      tags: [documents]
      summary: 업로드 문서 파싱·검수게이트·분류 단계별 진단
      description: |
        데모/관리자 콘솔용 read-only 분석 엔드포인트. 문서를 저장하지 않고
        추출 → 정규화·PII 마스킹 → 청킹 → 검수게이트 → 분류 결과를 한 번에 반환한다.
        parse에는 table_coverage, table_count, table_cell_count, warnings가 포함되어
        표 구조화 여부와 잠재 누락 경고를 확인할 수 있다.
        HWP/HWPX/Office/PDF 표 구조화와 Office 내부 이미지 OCR은 설치된 optional extractor에 따라
        수행되며, 미설치 또는 미검출 항목은 warnings로 노출된다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                return_evidence:
                  type: boolean
                  default: true
      responses:
        '200':
          description: 단계별 진단 결과
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: "파일 크기 초과 (MAX_UPLOAD_MB)" }
        '422': { description: "빈 파일 또는 파싱 실패" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    DocumentUploadResponse:
      type: object
      required: [filename, source_format, file_hash, file_size_bytes, extraction_method,
                 extraction_quality, ocr_used, char_count, chunk_count, persisted, warnings]
      properties:
        doc_id:
          type: string
          format: uuid
          nullable: true
          description: "DB 적재 성공 시 UUID, 미가용 시 null"
        filename: { type: string }
        source_format: { type: string, example: "pdf" }
        file_hash: { type: string, description: "SHA-256 hex" }
        file_size_bytes: { type: integer }
        extraction_method: { type: string, example: "pdfminer" }
        extraction_quality: { type: number, format: float, minimum: 0, maximum: 1 }
        ocr_used: { type: boolean }
        char_count: { type: integer }
        chunk_count: { type: integer }
        persisted: { type: boolean, description: "DB/storage 저장 성공 여부" }
        warnings:
          type: array
          items: { type: string }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계 식별용 오류 코드. 예시:
              KOIPA_VALIDATION, KOIPA_AUTH, KOIPA_NOT_FOUND,
              KOIPA_MODEL_NOT_READY, KOIPA_LLM_PROVIDER_FAIL,
              KOIPA_GPU_UNAVAILABLE, KOIPA_RATE_LIMIT, KOIPA_INTERNAL
        message: { type: string }
        request_id: { type: string, format: uuid }
        details:
          type: object
          additionalProperties: true
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InternalError:
      description: 서버 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...